backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
firebase-admin==6.4.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
google-generativeai==0.3.2
python-multipart==0.0.6
